# docs/swagger.yaml
definitions:
  models.Article:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  models.CreateArticle:
    properties:
      content:
        type: string
      title:
        maxLength: 256
        type: string
    required:
    - content
    - title
    type: object
  models.CreateList:
    properties:
      name:
        type: string
      next_page_key:
        type: string
    required:
    - name
    type: object
  models.CreatePage:
    properties:
      article_ids:
        items:
          type: integer
        type: array
      next_page_key:
        type: string
    required:
    - article_ids
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.List:
    properties:
      key:
        type: string
      name:
        type: string
      next_page_key:
        type: string
    required:
    - name
    type: object
  models.Page:
    properties:
      articles:
        items:
          $ref: '#/definitions/models.Article'
        type: array
      key:
        type: string
      next_page_key:
        type: string
    type: object
  models.UpdateArticle:
    properties:
      content:
        type: string
      title:
        maxLength: 256
        type: string
    required:
    - content
    - title
    type: object
  models.UpdateList:
    properties:
      name:
        type: string
      next_page_key:
        type: string
    required:
    - name
    type: object
  models.UpdatePage:
    properties:
      article_ids:
        items:
          type: integer
        type: array
      next_page_key:
        type: string
    required:
    - article_ids
    type: object
info:
  contact: {}
  description: This is the Backend Intern Homework api documentation
  title: Backend Intern Homework API
  version: "1.0"
paths:
  /articles:
    get:
      consumes:
      - application/json
      description: Get all articles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Article'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Create a article
      parameters:
      - description: Create article request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateArticle'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Article'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create article
      tags:
      - articles
  /articles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a article
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete article
      tags:
      - articles
    get:
      consumes:
      - application/json
      description: Find a article
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Article'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Find article
      tags:
      - articles
    patch:
      consumes:
      - application/json
      description: Update a article
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update article request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateArticle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Article'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update article
      tags:
      - articles
  /lists:
    get:
      consumes:
      - application/json
      description: Get all lists
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.List'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all lists
      tags:
      - lists
    post:
      consumes:
      - application/json
      description: Create a list
      parameters:
      - description: Create list request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateList'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.List'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create list
      tags:
      - lists
  /lists/{key}:
    delete:
      consumes:
      - application/json
      description: Delete a list
      parameters:
      - description: List Key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete list
      tags:
      - lists
    get:
      consumes:
      - application/json
      description: Get a list
      parameters:
      - description: List Key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.List'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get list
      tags:
      - lists
    patch:
      consumes:
      - application/json
      description: Update a list
      parameters:
      - description: List Key
        in: path
        name: key
        required: true
        type: string
      - description: Update list request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.List'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update list
      tags:
      - lists
  /pages:
    get:
      consumes:
      - application/json
      description: Get all pages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Page'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all pages
      tags:
      - pages
    post:
      consumes:
      - application/json
      description: Create a page
      parameters:
      - description: Create page request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreatePage'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create page
      tags:
      - pages
  /pages/{key}:
    delete:
      consumes:
      - application/json
      description: Delete a page
      parameters:
      - description: Page Key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete page
      tags:
      - pages
    get:
      consumes:
      - application/json
      description: Get a page
      parameters:
      - description: Page Key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get page
      tags:
      - pages
    patch:
      consumes:
      - application/json
      description: Update a page
      parameters:
      - description: Page Key
        in: path
        name: key
        required: true
        type: string
      - description: Update page request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update page
      tags:
      - pages
swagger: "2.0"